+++ src/main/java/uk/gov/service/notify/NotificationClientApi.java
@@ -7,7 +7,7 @@ public interface NotificationClientApi {
     /**
      * The sendEmail method will create an HTTPS POST request. A JWT token will be created and added as an Authorization header to the request.
      *
-     * @param templateId      Find templateId by clicking API info for the template you want to send
+     * @param templateId      The template id is visible on the template page in the application.
      * @param emailAddress    The email address
      * @param personalisation Map representing the placeholders for the template if any. For example, key=name value=Bob
      *                        Can be an empty map or null when the template does not require placeholders.
@@ -22,7 +22,7 @@ public interface NotificationClientApi {
     /**
      * The sendSms method will create an HTTPS POST request. A JWT token will be created and added as an Authorization header to the request.
      *
-     * @param templateId      Find templateId by clicking API info for the template you want to send
+     * @param templateId      The template id is visible from the template page in the application.
      * @param phoneNumber              The mobile phone number
      * @param personalisation Map representing the placeholders for the template if any. For example, key=name value=Bob
      *                        Can be an empty map or null when the template does not require placeholders.
@@ -58,5 +58,43 @@ public interface NotificationClientApi {
      */
     NotificationList getNotifications(String status, String notification_type, String reference, String olderThanId) throws NotificationClientException;
 
+    /**
+     * The getTemplateById returns a <code>Template</code> given the template id.
+     *
+     * @param templateId The template id is visible on the template page in the application.
+     * @return <code>Template</code>
+     * @throws NotificationClientException
+     */
+    Template getTemplateById(String templateId) throws NotificationClientException;
 
+    /**
+     * The getTemplateVersion returns a <code>Template</code> given the template id and version.
+     *
+     * @param templateId The template id is visible on the template page in the application.
+     * @param version The version of the template to return
+     * @return <code>Template</code>
+     * @throws NotificationClientException
+     */
+    Template getTemplateVersion(String templateId, int version) throws NotificationClientException;
+
+    /**
+     * Returns all the templates for your service. Filtered by template type if not null.
+     *
+     * @param templateType If templateType is not empty or null templates will be filtered by type.
+     *          Possible template types are email|sms|letter
+     * @return <code>TemplateList</code>
+     * @throws NotificationClientException
+     */
+    TemplateList getAllTemplates(String templateType) throws NotificationClientException;
+
+    /**
+     * The generateTemplatePreview returns a template with the placeholders replaced with the given personalisation.
+     *
+     * @param templateId The template id is visible from the template page in the application.
+     * @param personalisation Map representing the placeholders for the template if any. For example, key=name value=Bob
+     *                        Can be an empty map or null when the template does not require placeholders.
+     * @return <code>Template</code>
+     * @throws NotificationClientException
+     */
+    TemplatePreview generateTemplatePreview(String templateId, Map<String, String> personalisation) throws NotificationClientException;
 }